small refactor moving files to packages
