+++ src/main/java/com/oneencrypt/oneencrypt/central/services/fileservices/CreateFileNamePathService.java
@@ -1,4 +1,4 @@
-package com.oneencrypt.oneencrypt.central;
+package com.oneencrypt.oneencrypt.central.services.fileservices;
 
 import java.util.*;
